portreto/application/app_service/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dataclasses==0.6
Django==2.2.1
django-crispy-forms==1.7.2
django-stubs==0.12.1
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
djongo==1.2.32
idna==2.8
kazoo==2.6.1
mypy==0.701
mypy-extensions==0.4.1
Pillow==6.0.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
pymongo==3.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.2.4
typed-ast==1.3.5
typing-extensions==3.7.2
urllib3==1.24.3
